requirements.txt
appdirs==1.4.4
auxlib==0.0.43
setuptools>=40.3.0
beautifulsoup4==4.9.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
APScheduler==3.6.3
click==7.1.2
Deprecated==1.2.10
distlib==0.3.0
filelock==3.0.12
Flask==1.1.2
gcloud==0.18.3
gitdb==4.0.5
GitPython==3.1.3
google-api-core==1.17.0
google-api-python-client==1.8.4
google-auth==1.16.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==4.1.2
oauthlib==3.1.0
pipenv==2020.6.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.7
PyGithub==1.51
PyJWT==1.7.1
Pyrebase4==4.3.0
python-dotenv==0.13.0
python-jwt==2.0.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
smmap==3.0.4
soupsieve==2.0.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.21
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1